Experiment to add a unit test for SK21Drive
This adds a stub IMU class and testcase for the SK21Drive class. Currently, the test is trivial (merely to check that the system will build and run any test) but I see a system mutex exception when it exits. Given that the tests are run on every "Build Robot Code" call, this causes all builds to fail, even though the individual test actually passes. I'm checking this in on a branch while I continue investigation.